data-pipeline/pyproject.toml
[project]
name = "nba-shots"
version = "0.1.0"
description = "Visualize shot location and other aggregates for players, teams and the league."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pandas",
  "sqlalchemy",
  "mysqlclient",
  "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
# [project.scripts]
# hello = "uv_docker_example:hello"
# 
# [tool.uv]
# dev-dependencies = [
#     "ruff>=0.6.2",
#     "fastapi-cli>=0.0.5",
# ]
